requirements.txt
# ── Core server ────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
gunicorn>=22.0.0
python-multipart>=0.0.9

# ── LangGraph ─────────────────────────────────────────────────────────
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=0.2.0
python-dotenv>=1.0.0

# ── Gemini API (Primary LLM) ──────────────────────────────────────────
google-genai>=1.0.0

# ── Image processing ──────────────────────────────────────────────────
Pillow>=10.0.0
numpy>=1.24.0

# ── Data validation ───────────────────────────────────────────────────
pydantic>=2.7.0

# ── Optional: EasyOCR fallback (only if USE_EASYOCR=true) ─────────────
# easyocr>=1.7.1

# ── Optional: OpenAI compatibility ────────────────────────────────────
# openai>=1.35.0

# ── Groq API (Free fallback) ─────────────────────────────────────────
groq>=0.5.0
